SHA-115: Convert serviceworker to typescript (#266)
