+++ .env
@@ -1,3 +1,5 @@
 VERSION=2.3.5.0
 CHANGELOG_VERSION=2.3.3.0
+SENTRY_RELEASE=2.3.5.0
 GOOGLE_MEASUREMENT_ID="G-SN6JNMXND2"
+PROXY_URL=https://faas-nyc1-2ef2e6cc.doserverless.co/api/v1/web/fn-7932f4c9-dd5e-44e6-a067-5cbf1cf629d4/OpenAI_proxy/proxy
