+++ src/js/utils/getSyncedSetting.ts
@@ -1,4 +1,4 @@
-export async function getSyncedSetting(setting: string, defaultValue: string | undefined): Promise<string | undefined> {
+export async function getSyncedSetting(setting: string, defaultValue: string | boolean | undefined): Promise<string | undefined> {
   try {
     const result = await chrome.storage.sync.get(setting)
     const {[setting]: value = defaultValue} = result
