+++ tsconfig.json
@@ -15,7 +15,8 @@
     /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "lib": [
       "ES6",
-      "DOM"
+      "DOM",
+      "webworker",
     ],
     /* Specify a set of bundled library declaration files that describe the target runtime environment. */
     // "jsx": "preserve",                                /* Specify what JSX code is generated. */
